fixed review stuff.

+++ app.py
@@ -7,7 +7,14 @@ from flask_debugtoolbar import DebugToolbarExtension
 from sqlalchemy.exc import IntegrityError
 
 from forms import UserAddForm, LoginForm, MessageForm, CSRFProtectForm, UserEditForm
-from models import db, connect_db, User, Message
+from models import (
+    db,
+    connect_db,
+    User,
+    Message,
+    DEFAULT_HEADER_IMAGE_URL,
+    DEFAULT_IMAGE_URL,
+)
 
 load_dotenv()
 CURR_USER_KEY = "curr_user"
@@ -16,12 +23,13 @@ app = Flask(__name__)
 
 # Get DB_URI from environ variable (useful for production/testing) or,
 # if not set there, use development local db.
-app.config['SQLALCHEMY_DATABASE_URI'] = (
-    os.environ['DATABASE_URL'].replace("postgres://", "postgresql://"))
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-app.config['SQLALCHEMY_ECHO'] = False
-app.config['DEBUG_TB_INTERCEPT_REDIRECTS'] = True
-app.config['SECRET_KEY'] = os.environ['SECRET_KEY']
+app.config["SQLALCHEMY_DATABASE_URI"] = os.environ["DATABASE_URL"].replace(
+    "postgres://", "postgresql://"
+)
+app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
+app.config["SQLALCHEMY_ECHO"] = False
+app.config["DEBUG_TB_INTERCEPT_REDIRECTS"] = True
+app.config["SECRET_KEY"] = os.environ["SECRET_KEY"]
 toolbar = DebugToolbarExtension(app)
 
 connect_db(app)
@@ -42,6 +50,13 @@ def add_user_to_g():
         g.user = None
 
 
+@app.before_request
+def add_form_to_g():
+    """adds a CSRF form to Flask global."""
+
+    g.csrf_form = CSRFProtectForm()
+
+
 def do_login(user):
     """Log in user."""
 
@@ -54,13 +69,8 @@ def do_logout():
     if CURR_USER_KEY in session:
         del session[CURR_USER_KEY]
 
-@app.before_request
-def add_form_to_g():
-    """ adds a CSRF form to Flask global."""
 
-    g.csrf_form = CSRFProtectForm()
-
-@app.route('/signup', methods=["GET", "POST"])
+@app.route("/signup", methods=["GET", "POST"])
 def signup():
     """Handle user signup.
 
@@ -71,7 +81,7 @@ def signup():
     If the there already is a user with that username: flash message
     and re-present form.
     """
-   
+
     if CURR_USER_KEY in session:
         del session[CURR_USER_KEY]
     form = UserAddForm()
@@ -87,46 +97,45 @@ def signup():
             db.session.commit()
 
         except IntegrityError:
-            flash("Username already taken", 'danger')
-            return render_template('users/signup.html', form=form)
+            flash("Username already taken", "danger")
+            return render_template("users/signup.html", form=form)
 
         do_login(user)
 
         return redirect("/")
 
     else:
-        return render_template('users/signup.html', form=form)
+        return render_template("users/signup.html", form=form)
 
 
-@app.route('/login', methods=["GET", "POST"])
+@app.route("/login", methods=["GET", "POST"])
 def login():
     """Handle user login and redirect to homepage on success."""
 
     form = LoginForm()
 
     if form.validate_on_submit():
-        user = User.authenticate(
-            form.username.data,
-            form.password.data)
+        user = User.authenticate(form.username.data, form.password.data)
 
         if user:
             do_login(user)
             flash(f"Hello, {user.username}!", "success")
             return redirect("/")
 
-        flash("Invalid credentials.", 'danger')
+        flash("Invalid credentials.", "danger")
 
-    return render_template('users/login.html', form=form)
+    return render_template("users/login.html", form=form)
 
 
-@app.post('/logout')
+@app.post("/logout")
 def logout():
     """Handle logout of user and redirect to homepage."""
 
     if g.csrf_form.validate_on_submit():
         do_logout()
-        return redirect('/')
-
+        return redirect("/")
+    else:
+        raise Exception("GO THROUGH THE PROPER CHANNELS.")
 
     # IMPLEMENT THIS AND FIX BUG
     # DO NOT CHANGE METHOD ON ROUTE
@@ -135,7 +144,8 @@ def logout():
 ##############################################################################
 # General user routes:
 
-@app.get('/users')
+
+@app.get("/users")
 def list_users():
     """Page with listing of users.
 
@@ -146,18 +156,17 @@ def list_users():
         flash("Access unauthorized.", "danger")
         return redirect("/")
 
-    search = request.args.get('q')
+    search = request.args.get("q")
 
     if not search:
         users = User.query.all()
     else:
         users = User.query.filter(User.username.like(f"%{search}%")).all()
 
-    return render_template('users/index.html', 
-                            users=users)
+    return render_template("users/index.html", users=users)
 
 
-@app.get('/users/<int:user_id>')
+@app.get("/users/<int:user_id>")
 def show_user(user_id):
     """Show user profile."""
 
@@ -167,10 +176,10 @@ def show_user(user_id):
 
     user = User.query.get_or_404(user_id)
 
-    return render_template('users/show.html', user=user)
+    return render_template("users/show.html", user=user)
 
 
-@app.get('/users/<int:user_id>/following')
+@app.get("/users/<int:user_id>/following")
 def show_following(user_id):
     """Show list of people this user is following."""
 
@@ -179,24 +188,22 @@ def show_following(user_id):
         return redirect("/")
 
     user = User.query.get_or_404(user_id)
-    return render_template('users/following.html', 
-                            user=user)
+    return render_template("users/following.html", user=user)
 
 
-@app.get('/users/<int:user_id>/followers')
+@app.get("/users/<int:user_id>/followers")
 def show_followers(user_id):
     """Show list of followers of this user."""
-    
+
     if not g.user:
         flash("Access unauthorized.", "danger")
         return redirect("/")
 
     user = User.query.get_or_404(user_id)
-    return render_template('users/followers.html', 
-                            user=user)
+    return render_template("users/followers.html", user=user)
 
 
-@app.post('/users/follow/<int:follow_id>')
+@app.post("/users/follow/<int:follow_id>")
 def start_following(follow_id):
     """Add a follow for the currently-logged-in user.
 
@@ -214,7 +221,7 @@ def start_following(follow_id):
     return redirect(f"/users/{g.user.id}/following")
 
 
-@app.post('/users/stop-following/<int:follow_id>')
+@app.post("/users/stop-following/<int:follow_id>")
 def stop_following(follow_id):
     """Have currently-logged-in-user stop following this user.
 
@@ -232,68 +239,78 @@ def stop_following(follow_id):
     return redirect(f"/users/{g.user.id}/following")
 
 
-@app.route('/users/profile', methods=["GET", "POST"])
+@app.route("/users/profile", methods=["GET", "POST"])
 def profile():
     """Update profile for current user."""
 
-    form = UserEditForm()
+    user = g.user
+    form = UserEditForm(
+        username=user.username,
+        email=user.email,
+        image_url=user.image_url,
+        header_image_url=user.header_image_url,
+        bio=user.bio,
+    )
 
     if form.validate_on_submit():
-        user = g.user
+
         if User.authenticate(user.username, form.password.data):
-            user.username = form.username.data or user.username
-            user.email = form.email.data or user.email
-            user.image_url = form.image_url.data or user.image_url
-            user.header_image_url = form.header_image_url.data or user.header_image_url
-            user.bio = form.bio.data or user.bio
+            user.username = form.username.data
+            user.email = form.email.data
+            user.image_url = form.image_url.data or DEFAULT_IMAGE_URL
+            user.header_image_url = (
+                form.header_image_url.data or DEFAULT_HEADER_IMAGE_URL
+            )
+            user.bio = form.bio.data
             db.session.commit()
 
-            return redirect(f'/users/{user.id}')
+            return redirect(f"/users/{user.id}")
         else:
-            flash('Bad password.')
+            flash("Bad password.")
             username = form.username.data
             email = form.email.data
             image_url = form.image_url.data
             header_image_url = form.header_image_url.data
             bio = form.bio.data
-            form = UserEditForm(username=username,
-                                email=email,
-                                image_url=image_url,
-                                header_image_url=header_image_url,
-                                bio=bio)
-            return render_template('users/edit.html', 
-                                form=form)
-            
-        
-    else:    
-        return render_template('users/edit.html', 
-                                form=form)
+            form = UserEditForm(
+                username=username,
+                email=email,
+                image_url=image_url,
+                header_image_url=header_image_url,
+                bio=bio,
+            )
+            return render_template("users/edit.html", form=form)
+
+    else:
+        return render_template("users/edit.html", form=form)
 
 
-
-@app.post('/users/delete')
+@app.post("/users/delete")
 def delete_user():
     """Delete user.
 
     Redirect to signup page.
     """
+    if g.csrf_form.validate_on_submit():
+        if not g.user:
+            flash("Access unauthorized.", "danger")
+            return redirect("/")
 
-    if not g.user:
-        flash("Access unauthorized.", "danger")
-        return redirect("/")
+        do_logout()
 
-    do_logout()
+        db.session.delete(g.user)
+        db.session.commit()
 
-    db.session.delete(g.user)
-    db.session.commit()
-
-    return redirect("/signup")
+        return redirect("/signup")
+    else:
+        raise Exception("Get Validated bruh")
 
 
 ##############################################################################
 # Messages routes:
 
-@app.route('/messages/new', methods=["GET", "POST"])
+
+@app.route("/messages/new", methods=["GET", "POST"])
 def add_message():
     """Add a message:
 
@@ -313,10 +330,10 @@ def add_message():
 
         return redirect(f"/users/{g.user.id}")
 
-    return render_template('messages/new.html', form=form)
+    return render_template("messages/new.html", form=form)
 
 
-@app.get('/messages/<int:message_id>')
+@app.get("/messages/<int:message_id>")
 def show_message(message_id):
     """Show a message."""
 
@@ -325,10 +342,10 @@ def show_message(message_id):
         return redirect("/")
 
     msg = Message.query.get_or_404(message_id)
-    return render_template('messages/show.html', message=msg)
+    return render_template("messages/show.html", message=msg)
 
 
-@app.post('/messages/<int:message_id>/delete')
+@app.post("/messages/<int:message_id>/delete")
 def delete_message(message_id):
     """Delete a message.
 
@@ -351,7 +368,7 @@ def delete_message(message_id):
 # Homepage and error pages
 
 
-@app.get('/')
+@app.get("/")
 def homepage():
     """Show homepage:
 
@@ -361,17 +378,19 @@ def homepage():
     if g.user:
         following = [u.id for u in g.user.following]
 
-        messages = (Message
-                    .query
-                    .filter(Message.user_id == g.user.id or Message.user_id.in_(following))
-                    .order_by(Message.timestamp.desc())
-                    .limit(100)
-                    .all())
+        messages = (
+            Message.query.filter(
+                Message.user_id == g.user.id or Message.user_id.in_(following)
+            )
+            .order_by(Message.timestamp.desc())
+            .limit(100)
+            .all()
+        )
 
-        return render_template('home.html', messages=messages)
+        return render_template("home.html", messages=messages)
 
     else:
-        return render_template('home-anon.html')
+        return render_template("home-anon.html")
 
 
 ##############################################################################
@@ -380,6 +399,7 @@ def homepage():
 #   handled elsewhere)
 #
 # https://stackoverflow.com/questions/34066804/disabling-caching-in-flask
+
 
 @app.after_request
 def add_header(response):
